+++ R/scales.R
@@ -27,6 +27,7 @@ scale_x_break <- function(breaks, scales="fixed", ticklabels=NULL, expand=TRUE, 
 ##' @title scale_y_break
 ##' @rdname scale_break
 ##' @export
+##' @import yulab.utils
 ##' @examples
 ##' require(ggplot2 )
 ##' set.seed(2019-01-19)
@@ -48,9 +49,12 @@ scale_break <- function(axis, breaks, scales, ticklabels=NULL, expand=TRUE, spac
               class = "ggbreak_params")
 }
 
-#' @title scale_wrap
-#'
+
 #' This scale function wraps a 'gg' plot over multiple rows to make plots with long x axes easier to read.
+#'
+#' 
+#' @title scale-wrap
+#' @rdname scale-wrap
 #' @param n the number of subplot pieces.
 #' @return gg object
 #' @export